Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -66,10 +66,13 @@ router.put(
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({_id: req.params.id},  post)
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId},  post)
     .then(result => {
-      res.status(200).json({
-        message: 'Post updated successfully!'})
+      if (result.nModified > 0) {
+        res.status(200).json({message: 'Post updated successfully!'});
+      } else {
+        res.status(401).json({message: "Not authorized!"});
+      }
     })
 });
 
@@ -109,9 +112,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then(result => {
-      res.status(200).json({message: "Post deleted!"});
+      if (result.n > 0) {
+        res.status(200).json({message: "Post deleted!"});
+      } else {
+        res.status(401).json({message: "Not authorized!"});
+      }
     });
 });
